Creates BST
Added factory functions for both creating nodes, and for creating the
BST itself. Added a function that builds the tree given an array, as
well as functions for inserting and deleting nodes from the tree. Added
a minValue function as well to help with the deletion function. Has a
temporary script just for testing out if everything is working correctly